refactor: Calculate -> CalculatorFunction 으로 명사화
