--- src/main/java/calculator/Calculate.java
@@ -1,19 +1,0 @@
-package calculator;
-
-public class Calculate {
-    public double addition(double firstNumber, double secondNumber) {
-        return firstNumber + secondNumber;
-    }
-
-    public double subtraction(double firstNumber, double secondNumber) {
-        return firstNumber - secondNumber;
-    }
-
-    public double multiplication(double firstNumber, double secondNumber) {
-        return firstNumber * secondNumber;
-    }
-
-    public double division(double firstNumber, double secondNumber) {
-        return firstNumber / secondNumber;
-    }
-}
