+++ src/main/java/calculator/CalculatorFunction.java
@@ -1,0 +1,19 @@
+package calculator;
+
+public class CalculatorFunction {
+    public double addNumber(double firstNumber, double secondNumber) {
+        return firstNumber + secondNumber;
+    }
+
+    public double subtractNumber(double firstNumber, double secondNumber) {
+        return firstNumber - secondNumber;
+    }
+
+    public double multipleNumber(double firstNumber, double secondNumber) {
+        return firstNumber * secondNumber;
+    }
+
+    public double divideNumber(double firstNumber, double secondNumber) {
+        return firstNumber / secondNumber;
+    }
+}
